373506498: Fix false positives in lint API checks
This CL fixes a corner case in lint's SDK_INT arithmetic.
When lint is computing version constraints (SDK_INT > X) in the
source code, it sometimes needs to negate the constraint; for example,
when computing the cumulative constraint in a Kotlin when statement
like this:
   when {
      SDK_INT >= 34 -> ...
      // when we get here, we negate the above constraint, so
      // here we know SDK_INT < 34
      SDK_INT >= 30 -> ... // now we know SDK_INT >= 30 && SDK_INT < 34
      // now we know SDK_INT < 30, and so on.

There's a similar situation for if/else blocks.

There was a problem here; for an if block like

      if (SDK_INT >= 34 && someCondition() {

here we know that SDK_INT >= 34 in the then block. But we can NOT
conclude that it's < 34 in the else clause.

To solve this, we now track whether an ApiConstraint is negatable.
When we combine ApiConstraint in &&, we mark it as non-negatable.

Fixes: 373506498
Test: Many existing, and a few updated
Change-Id: I99bb335833abfcb261996d8ce7284273b7e7e244