Create Proxy for AGP extension for integration tests
This allows configuring the test project DSL with the
same interface that implements AGP's DSL

Order of calls is important. For instance of the order of calls
to compileSdk, compileSdkPreview and compileSdkAddon matters.

So we cannot just record values for each property in a maps.
We have to record call events, in certain order, and replay
them in the same order.

The mechanism uses a DslContentHolder class responsible for
holding the events. This is handled by the Proxy class that
just delegate events that it understands to the holder.
A Chained content holder is used for chained calls (e.g.
android.androidResources.enable = true), in which case
only the original holder records the event.
Actual nested block uses its owner sub-holder that is
considered an event.

List/Set are also handled via the main holder. There is
a custom implementations that just sends even to the parent
holder.

Bug: N/A
Test: Units tests for holder class and the proxy.
Change-Id: Ie2d689a0bc6c2251f5115df24f6973773c79168c