GradleRule: Support for buildtypes / flavors.
Implements proxy for the NamedDomainObjectContainerProxy to support
basic methods like create and named.

One issue this runs into is the fact that Java Proxy cannot be created
for ProductFlavor due to b/377720361. Therefore in order to implement
a proxy for this interface, we use ByteBuddy which offers a similar
mechanism of method interceptor.

Bug: N/A
Test: converted new tests to validate implementation
Change-Id: Ie04edd50101c99e74b6a61225dce0acd0371f2c3