Remove use of stub_apk in GMD-FTL
A stub_apk was used as a workaround solution for
the FTL TestLab backend requring two different APKs
to run instrument test. When you run a self-instrumented
test (e.g. androidTest in the library module), there
is only a test apk and no app apk.

This restriction from the FTL TestLab API is removed and
it accepts a single APK to be set for both app and test
APK fields.

Bug: 373667192
Test: unit tests and manual e2e validation
Change-Id: I92175c63bc163839580125c29de131987a393992