Update to API 35 and fix appop permissions (b/375352607)
This CL fixes a number of issues around platform metadata:

First, update lint platform metadata to Android 15.  Lint needs some
information which isn't packaged with the SDK, such as a mapping from
Context constants to the corresponding service class implementation.
This metadata needs to be updated manually from time to time; this CL
updates it for Android 15. I also updated some of the code which
generates this test data such that they can operate on released SDK
artifacts instead of pointing to source files (since it's sometimes
tedious to find the right branch for this.)

Second, and the trigger for this work, is a fix for
375352607: SCHEDULE_EXACT_ALARM is no longer only for system apps

The root issue here is that we were not correctly handling permissions
marked as "appops". This is one of the oldest lint checks, and when
appops permissions was introduced, this lint check was not updated, and
I've incorrectly spot fixed a few other reports of these other
permissions. The correct fix is to not treat these permissions as system
only, and that also gets rid of the individual special-cases.

Finally, update the code which already filtered on maxSdkVersion on the
permission tag to also ignore permissions removed before the
minSdkVersion.

Fixes: 375352607
Test: Unit tests included
Change-Id: Ie6268d63c1ee92618890dac102d2ebd39596b462