GradleRule: Kotlin support.
This change adds kotlin support by allowing
setting kotlin options. This is done by another DSL
support in a kotlin block to represent the top level
extension exposed by the built-in support for kotlin.

Originally, the plan was to use Kotlin's own extension
interfaces, however, it turned out to be impossible. The
Kotlin DSL is not exposed as pure interfaces, but as a
mix of interfaces and classes, so Java Proxy does not
work with them. ByteBuddy also does not work due to
some complex issue (the error indicated native methods!)

In the end, this change forks part of the interface but
reuse some in order to expose the currently used DSL
options. Everything is still going through the proxy
so the interfaces are the only thing we need to maintain.

All this currently requires the new built-in support. We
may add support for not standalone KGP plugin, or not.
We'll see.

Another change has to do with plugin versions. This change
adds support for per-project plugin versions, and the
generated build files handle properly when separate
projects use different versions or whether they all use
the same version.

Bug: N/A
Test: new basic test for Kotlin.
Change-Id: I23245d023aac436d05d8969ff2e5af7c13602408