Further improvements to the SamDetector
This CL generalizes the SAM detector such that for remove calls, it
doesn't only flag the Handler.removeCallbacks instances, it flags *any*
call to a method that starts with the name "remove" and passes in
something not referenceable - e.g. SAM conversions and constructor
calls.

Second, if the instance passed in is a lambda variable, it offers a
quickfix to construct an instance matching the interface type (e.g. for
`val runner = { something() }` it will convert it into `val runner =
Runnable { something() }`.

Third, the SAM detector was actually off by default! This CL (and the
corresponding on the IDE plugin side) turns it on by default.

Fourth, it now always makes sure (using the analysis API) that we're
really dealing with a SAM conversion; this fixes a few false positives
in the AndroidX repository (and probably elsewhere).

Fifth, dial back the heuristics which was guessing whether storing a
lambda inside a method is a problem. This was the case for listeners,
but there are various counter examples (such as the new testWear unit
test), so instead, we limit these checks to whether the lambda is
compared by instance or passed on to a removal method.

Finally, it updates the quickfix machinery for using regular expressions
to update code such that it allows line terminators to match dot (so you
can easily have a quickfix span multiple lines).

Bug: 376498180
Test: Unit tests included
Change-Id: I18f251d2d10d5dfc53617eb56a2802af1c2d50af