GradleRule: Support for AiPack
Although AiPack is part of AGP, because they
don't share a root extension with the "core" AGP
plugins, the implementation of the AI Pack support
does not extend the AndroidProject[Definition]
interfaces.

This triggered some changes around inheritance
of the 3 types of project (definition, project,
reversible projects) which was inconsistent.
Recent introductions of BaseGradleProject
was not updated everywhere and places that
still manipulated all project types referred
to them as GradleProject instead of BaseGradleProject.
This commit fixes this, and pulls some feature
down to the base interface (reconfigure) which
also trigger cleaning some of the type
parameters.

In a later CL we'll probably rename [Base]GradleProject.

This commit also adds support for testing
app bundle (aab) from an android application

Bug: N/A
Test: existing tests
Change-Id: Ide5748e3dca14ba6a882c0ba22efd610c88b39c0